baba_prospector/requirements.txt
python-telegram-bot==20.7
groq==0.9.0
gspread==6.1.2
google-auth==2.29.0
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.1
